Remove duplicate example, point to examples on docs index page (#84)
This PR is based on my own developer experience, of not knowing how to
make backon work with an async call that takes variable arguments.

To avoid having duplicate examples is various places, we rather point to
the documentation index page, where the main collection of examples
seems to be placed. This way hopefully developers can find the examples
they need to find, more easily.