Removing unnecessary optional feature selectors
Sorry for nitpicking, I was wondering what the question marks mean in the feature deps and after learning about their meaning it occured to me that they are actually not necessary here, since the tokio respectively gloo-timers crate is added as a dependency in the same feature. Hope I understood it correctly, happy to learn more about this syntax.