jitter: Make jitter range to current_delay instead of min_delay (#184)
Currently in exponential and fibonacci the jitter adds a value between 0
and min_delay, which might be very small when compared to the end of the
max_delay. So when reaching max_delay, and when many different endpoints
use the backoff algorithm, they will hit the backends in waves

This changes it to range from 0 to current_delay, so the jitter is
always a relevant size compared to the delays.

If you prefer this to be a different parameter, let me know an I will
adjust.