feat: Implement retry for FnMut() -> impl Future<Output=Option<Result<T, E>>>
This changes the way errors are handled via an intermediary trait
`BorrowError` which the function's return type must implement.

If a value of this type contains an error that should influence
the retry behavior, it will be returned by `as_error` function
in `BorrowError` trait implementation.

This allows to use any type as return function, including `Option<Result<T, E>>`.