feat: Add retry type for use with results
This commit allows using functions that operate on a `&Result<T, E>`
rather than the error type `E`. This may be useful in scenarios where,
when retries are exhausted, we want to return an `Ok(response)` from the
inner function being retried (along with an error status code), rather
than having the inner function return an `Err(_)`. This can be
useful in a gateway use case where we want to pass through responses
rather than map a HTTP response to a Rust Error, then back to an HTTP
response type